Feat : useSearch custom hook
