+++ components/Search/Search.tsx
@@ -1,28 +1,10 @@
 import Input from "../common/Input/Input";
-import { useCallback, useMemo } from "react";
-import { useSearchParams } from "next/navigation";
-import debounce from "lodash.debounce";
 import search from "@/public/icons/search.svg";
+import useSearch from "./useSearch";
 
-interface SearchProps {
-  onChange: (keyword: string) => void;
-}
+export default function Search() {
+  const { defaultValue, handleSearchChange } = useSearch();
 
-export default function Search({ onChange }: SearchProps) {
-  const searchParams = useSearchParams();
-
-  const debouncedKeyword = useMemo(
-    () => debounce((search: string) => onChange(search), 300),
-    []
-  );
-
-  const handleSearchChange = useCallback(
-    (e: React.ChangeEvent<HTMLInputElement>) => {
-      const search = e.target.value;
-      debouncedKeyword(search);
-    },
-    [debouncedKeyword]
-  );
   return (
     <>
       <Input
@@ -30,7 +12,7 @@ export default function Search({ onChange }: SearchProps) {
         leftSlot={search}
         slotSize={24}
         placeholder="검색할 상품을 입력해주세요"
-        defaultValue={searchParams.get("keyword")?.toString()}
+        defaultValue={defaultValue}
       />
     </>
   );
